Fix: RouteRepositoryImpl 클래스 getReservationSeats 함수 수정 select에 count() 함수는 long type을 반환하는데 좌석수니까 int로 변경. RouteDetailDto, UserReservedRouteDto도 int로 수정

+++ BE/src/main/java/com/lisade/togeduck/dto/response/RouteDetailDto.java
@@ -17,7 +17,7 @@ public class RouteDetailDto {
     String destination;
     LocalTime expectedAt;
     Integer totalSeats;
-    Long reservedSeats;
+    Integer reservedSeats;
     Integer cost;
 
 }
